requirements_test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements_test.txt requirements_test.in
#
-e git+https://github.com/stuaxo/django-admin-ip-restrictor.git@feature/block_by_url#egg=django-admin-ip-restrictor
-e git+https://github.com/infoportugal/wagtail-modeltranslation.git@v0.9.0#egg=wagtail_modeltranslation
ansicolor==0.2.6
attrs==17.4.0             # via pytest
beautifulsoup4==4.6.0     # via wagtail
bleach==2.1.3             # via wagtail-markdown
boto3==1.7.10
botocore==1.10.10
cachetools==2.0.1         # via zenpy
certifi==2018.4.16        # via requests
chardet==3.0.4            # via requests
codecov==2.0.15
coverage==4.5.1           # via codecov, pytest-cov
directory-validators==4.5.1
dj-database-url==0.4.2
django-clear-cache==0.3
django-crispy-forms==1.7.1
django-ipware==1.1.6
django-modelcluster==4.1  # via wagtail
django-modeltranslation==0.12.2
django-redis==4.8.0
django-storages==1.6.5
django-taggit==0.22.2     # via wagtail
django-treebeard==4.3     # via wagtail
django==1.11.11
django_recaptcha==1.2.0
djangorestframework==3.8.2  # via wagtail
docutils==0.14            # via botocore
draftjs-exporter==2.0.0   # via wagtail
factory-boy==2.7.0
fake-factory==0.7.4       # via factory-boy, wagtail-factories
faker==0.8.15             # via wagtail-factories
flake8==3.5.0
gunicorn==19.7.1
html5lib==0.999999999     # via bleach, wagtail
idna==2.6                 # via requests
jmespath==0.9.3           # via boto3, botocore
markdown==2.6.11          # via wagtail-markdown
mccabe==0.6.1             # via flake8
more-itertools==4.1.0     # via pytest
olefile==0.44             # via directory-validators, pillow
phonenumberslite==7.7.4   # via directory-validators
pillow==4.2.1             # via directory-validators, wagtail
pluggy==0.6.0             # via pytest
psycopg2==2.7.3.2
py==1.5.3                 # via pytest
pycodestyle==2.3.1        # via flake8
pyflakes==1.6.0           # via flake8
pytest-cov==2.5.1
pytest-django==3.2.1
pytest-sugar==0.9.1
pytest==3.5.1
python-dateutil==2.7.2    # via botocore, fake-factory, faker, zenpy
pytz==2017.2
raven==6.5.0
redis==2.10.6             # via django-redis
requests==2.18.4          # via codecov, wagtail, zenpy
s3transfer==0.1.13        # via boto3
six==1.11.0               # via bleach, fake-factory, faker, html5lib, more-itertools, pytest, python-dateutil
termcolor==1.1.0          # via pytest-sugar
text-unidecode==1.2       # via faker
unidecode==0.4.21         # via wagtail
urllib3==1.22             # via requests
wagtail-factories==1.0.0
wagtail-markdown==0.5a3
wagtail==2.0.1
waitress==1.1.0
webencodings==0.5.1       # via html5lib
whitenoise==3.3.1
willow==1.1               # via wagtail
zenpy==1.1.10
